Fixed: Github tarball instability.
When a repo uses .gitattributes export-subst
to stamp a tarball date into the tarball,
and github throws away it's cached tarball,
the tarball's hash will change upon redownload,
making reproducibility = no.

This detects the situation when computing the hash
in our 'trust-on-first-use' basis,
and replaces it with fetchgit, which does not suffer the problem.

This patch also introduced hash prefetching for fetchgit,
just like it's done for fetchFromGithub.