Replace additional_mkpython_arguments with additional_mkpython_arguments_func
We had been '// merging additional_mkpython_arguments_func with what
anysnake2 itself set on the mkPython call.
This would of course overwrite for example the _.rpy2 adjustments.

We now take a function in additional_mkpython_arguments_func,
that takes an 'old' attrSet, and it's result is // merged into a new
one.